+++ api/database.json.example
@@ -1,125 +1,201 @@
 {
-    "users": [
-          {
-      "id": 1,
-      "username": "Sophia Hoffman",
-      "email": "sophiahoffman@google.com"
-    },
+  "users": [
     {
-      "id": 2,
-      "username": "Rob Hoffman",
-      "email": "robhoffman@google.com"
-    },
-    {
-      "id": 3,
-      "username": "Moet Hoffman",
-      "email": "moethoffman@google.com"
-    },
-    {
-      "id": 4,
-      "username": "Chandon Hoffman",
-      "email": "chandonhoffman@google.com"
-    }
-    ],
-    "articles": [],
-    "messages": [
-          {
-      "id": 1,
-      "userId": 1,
-      "message": "I can make it work",
-      "timestamp": "2012-04-23T18:25:43.511Z"
-    },
-    {
-      "id": 2,
-      "userId": 1,
-      "message": "Noooo.... THIS is the coolest thing ever",
-      "timestamp": "2013-04-23T18:25:43.511Z"
-    },
-    {
-      "id": 3,
-      "userId": 2,
-      "message": "I say, old chap; can we do this again?",
-      "timestamp": "2014-04-23T18:25:43.511Z"
-    },
-    {
-      "id": 4,
-      "userId": 2,
-      "message": "Let's give em something to talk about",
-      "timestamp": "2016-04-23T18:25:43.511Z"
-    },
-    {
-      "userId": 1,
-      "message": "I love this!!!",
-      "timestamp": 1573502178009,
-      "id": 5
-    },
-    {
-      "userId": 1,
-      "message": "This is great",
-      "timestamp": 1573503225202,
-      "id": 6
-    },
-    {
-      "userId": 1,
-      "message": "hello hello hello",
-      "timestamp": 1573510304481,
-      "id": 7
-    },
-    {
-      "userId": 1,
-      "message": "I think this is awesome",
-      "timestamp": 1573510573473,
-      "id": 8
-    }
-    ],
-    "events": [],
-    "tasks": [
-       {
-      "id": 1,
-      "userId": 1,
-      "task": "testing code",
-      "compDate": "2019-12-02",
-      "completed": false
-    },
-    {
-      "userId": "1",
-      "task": "testing code take 2",
-      "compDate": "2019-03-18",
-      "completed": false,
-      "id": 2
-    },
-    {
-      "userId": "1",
-      "task": "testing code take 3",
-      "compDate": "2018-11-19",
-      "completed": false,
-      "id": 3
-    },
-    {
-      "userId": "1",
-      "task": "testing code",
-      "compDate": "2019-11-13",
-      "completed": false,
-      "id": 4
-    },
-    {
-      "userId": "1",
-      "task": "testing code",
-      "compDate": "2019-11-08",
-      "completed": false,
-      "id": 5
-    }
-    ],
-    "friends": [
-          {
-      "userId": 1,
-      "friendName": "happy happy",
+      "username": "hank",
+      "email": "hank@hank.com",
       "id": 1
     },
     {
-      "userId": 1,
-      "friendName": "john doe",
+      "username": "ryan",
+      "email": "ryan@ryan.com",
       "id": 2
+    },
+    {
+      "username": "steve",
+      "email": "steve@steve.com",
+      "id": 3
+    },
+    {
+      "username": "john",
+      "email": "john@john.com",
+      "id": 4
+    },
+    {
+      "username": "sophia",
+      "email": "sophia@sophia.com",
+      "id": 5
+    },
+    {
+      "username": "tom",
+      "email": "tom@tom.com",
+      "id": 6
     }
-    ]
+  ],
+  "articles": [
+    {
+      "userId": 5,
+      "articleName": "User Article",
+      "synopsis": "my article",
+      "articleUrl": "http://www.me.com",
+      "timestamp": 1573766140262,
+      "id": 1
+    },
+    {
+      "userId": 2,
+      "articleName": "FRIEND Article",
+      "synopsis": "my article",
+      "articleUrl": "http://www.me.com",
+      "timestamp": 1573766140262,
+      "id": 1
+    }
+  ],
+  "messages": [
+    {
+      "id": 1,
+      "userId": 1,
+      "message": "Howdy partner",
+      "timestamp": "1573583190113"
+    },
+    {
+      "id": 2,
+      "userId": 1,
+      "message": "hello",
+      "timestamp": "1573676963421"
+    },
+    {
+      "id": 3,
+      "userId": 2,
+      "message": "howdy",
+      "timestamp": "1573583190114"
+    },
+    {
+      "userId": 5,
+      "message": "it is me",
+      "timestamp": 1573766244932,
+      "id": 4
+    }
+  ],
+  "events": [
+    {
+      "userId": 1,
+      "name": "Party",
+      "date": "2019-11-28",
+      "location": "here",
+      "id": 1
+    },
+    {
+      "userId": 2,
+      "name": "party",
+      "date": "2019-12-31",
+      "location": "somewhere",
+      "id": 2
+    },
+    {
+      "userId": 5,
+      "name": "testing",
+      "date": "2013-01-02",
+      "location": "hackery",
+      "id": 4
+    },
+    {
+      "userId": 5,
+      "name": "hello",
+      "date": "2020-01-02",
+      "location": "nashvegas",
+      "id": 5
+    }
+  ],
+  "tasks": [
+    {
+      "userId": 1,
+      "task": "have fun",
+      "compDate": "2019-11-08",
+      "completed": false,
+      "id": 2
+    },
+    {
+      "userId": 1,
+      "task": "Finish phase two of tasks",
+      "compDate": "2019-11-08",
+      "completed": false,
+      "id": 8
+    },
+    {
+      "userId": 1,
+      "task": "tasks phase 2",
+      "compDate": "2019-11-08",
+      "completed": false,
+      "id": 9
+    },
+    {
+      "userId": 1,
+      "task": "Finish phase two of tasks",
+      "compDate": "2019-11-08",
+      "completed": false,
+      "id": 10
+    },
+    {
+      "id": 11,
+      "userId": 1,
+      "task": "this functional checkbox",
+      "compDate": "2019-11-07",
+      "completed": false
+    },
+    {
+      "userId": 1,
+      "task": "howdy",
+      "compDate": "2018-11-11",
+      "completed": true,
+      "id": 13
+    },
+    {
+      "userId": 1,
+      "task": "this should be on api but not on page",
+      "compDate": "2014-03-03",
+      "completed": true,
+      "id": 14
+    },
+    {
+      "userId": 1,
+      "task": "howdy",
+      "compDate": "1212-12-12",
+      "completed": true,
+      "id": 15
+    },
+    {
+      "userId": 1,
+      "task": "yoohoo",
+      "compDate": "1208-12-12",
+      "completed": true,
+      "id": 16
+    },
+    {
+      "userId": 1,
+      "task": "wtf",
+      "compDate": "1999-12-12",
+      "completed": true,
+      "id": 17
+    },
+    {
+      "userId": 1,
+      "task": "thought this would be easy",
+      "compDate": "2017-11-11",
+      "completed": true,
+      "id": 18
+    },
+    {
+      "userId": 1,
+      "task": "Finish delete button",
+      "compDate": "2019-11-30",
+      "completed": false,
+      "id": 19
+    }
+  ],
+  "friends": [
+    {
+      "currentUserId": 5,
+      "userId": 2,
+      "id": 1
+    }
+  ]
 }
